requirements_graphrag.txt
# Requirements for using neo4j_graphrag and OpenAI
openai>=1.0.0,<2.0.0
neo4j>=5.17.0,<6.0.0
neo4j-graphrag==1.0.0
pydantic>=2.0.0
python-dotenv
jinja2>=3.0.0
streamlit>=1.31.0
